fix trpc circle dependencies
