+++ src/server/api/services/permisos/permisos.service.ts
@@ -1,17 +1,7 @@
 import { protectedProcedure } from "../../trpc";
-import { getUsuarioYPermisos, verificarPermisoUsuario } from "../../repositories/permisos/permisos.repository";
+import { getUsuarioYPermisos } from "../../repositories/permisos/permisos.repository";
 import { permisosSchema } from "@/shared/filters/permisos-filter";
-import { type SgeNombre, type PrismaClient } from "@prisma/client";
-
-export const tienePermiso = async (ctx: { db: PrismaClient }, permisos: SgeNombre[], userId: string) => {
-  if (permisos.length === 0) {
-    return true;
-  }
-
-  const tienePermiso = await verificarPermisoUsuario(ctx, userId, permisos);
-
-  return tienePermiso;
-};
+import { tienePermiso } from "./permisos.helper";
 
 export const getUsuarioPermiso = protectedProcedure.query(async ({ ctx }) => {
   const userId = ctx.session.user.id;
